requirements.txt
pandas==0.25.1
discord.py==1.3.3
discord==1.0.1
python-dotenv==0.13.0
